started moving over to a controller approach
