+++ Auxilium/RoutedPages/api/v2/generate-pdf.php
@@ -1,24 +1,12 @@
 <?php
 
 use Auxilium\APITools;
+use Auxilium\Auxilium\API\Controllers\PDFController;
 use Auxilium\Helpers\PDF\PDFGeneration;
 use Auxilium\Utilities\URIUtilities;
 
 require_once __DIR__ . '/../../../vendor/autoload.php';
 require_once __DIR__ . '/../../../Configuration/Configuration/Environment.php';
 
-$at = APITools::get_instance();
-$at->requireLogin();
-
-$uri = new URIUtilities();
-$type = $uri->getURIComponents()[4];
-$uuid = $uri->getURIComponents()[5];
-
-switch($type)
-{
-    case "case":
-        PDFGeneration::GenerateCaseOverviewPage(caseID: "{{$uuid}}")->Render();
-    default:
-        $at->setErrorText(value: "PDF Type not recognized");
-        $at->output();
-}
+$controller = new PDFController();
+$controller->GeneratePDF();
